+++ src/Commands/WmModelCommands.php
@@ -84,6 +84,10 @@ class WmModelCommands extends DrushCommands implements SiteAliasManagerAwareInte
         'show-machine-names' => InputOption::VALUE_OPTIONAL,
     ]): void
     {
+        if (empty($options['output-module'])) {
+            throw new \InvalidArgumentException('You must specify an output module through --output-module or through configuration.');
+        }
+
         if (!$bundle) {
             $this->input->setArgument('bundle', $bundle = $this->askBundle());
         }