+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Remove implicit eck dependency
+- Validate the presence of the output-module option
 
 ## [1.13.1] - 2022-01-05
 ### Fixed